Going XDG, with active opt-in by the user
This is a proof of concept.

The basic idea is: If the user creates all the needed
(currently two) directories, then newsbeuter will go XDG,
otherwise it will go for the classic ~/.newsbeuter.  This
should make testing simple while keeping everyone else
happy.

To aid in moving to XDG, the startup messages contain
notices for the missing directory and the resulting
utilisation of the classic layout.

Current layout:
- "config" and "urls" in ~/.config/newsbeuter
- Everything else in ~/.local/share/newsbeuter