Fix crash when a GUID is lost while browsing command is running
Fix an abnormal termination which happens when a GUID is lost due
to the item being updated or deleted by the fetch thread when the
web browser command is running in the main thread. When the browser
finishes, newsbeuter attempts to find the item being read and calls
abort() on failure.

This crash is more frequent when using text-mode browsers (e.g. w3m)
which, unlike the graphical ones, do not return immediately.

This fix is more a hack than a proper one. The problem was exposed
when a lock was added to method rss_feed::get_item_by_guid_unlocked()
in commit d71336f48be10df737b811edc7796d6a9a88498e. After that,
failure to find the item resulted in abort() being called. This fix
brings back the dummy item introduced in commit
8dd8a153f8fd5b2fb86eb74e842ffe621c0ea4ba, avoiding the crash but,
possibly, showing an empty article.