Add --home-dir and NEWSBEUTER_HOME env variable
The only way to change home directories until now were the -u, -c and -C
arguments, and even with these you weren't able to change the history
files. This commit implements the --home-dir argument which switches all
files to the directory provided.

Additionally it adds support for the NEWSBEUTER_HOME environment
variable that behaves exactly as --home-dir.

With this commit Newsbeuter and Podbeuter will look for their files in
this order:

1. -u, -c and -C arguments
2. --home-dir
3. NEWSBEUTER_HOME environment variable
4. XDG directories
5. $HOME/.newsbeuter