Revert "Link Newsbeuter with libintl"
This reverts commit 72cd358691932591bda4b7ddb48cbe80238b95f6, which was
created in a haste and without full understanding of the situation.

Not that I understand anything now…

After a discussion at #musl on Freenode it seems to me that the problem
was specific to current situation in Alpine Linux and doesn't affect
anyone else. I'm reverting the commit since it broke all Linux builds.