Open multiple articles in browser (#380)
* Open multiple articles in browser

This adds commands to open multiple unread articles from the same
feed in the browser. These commands are named
open-all-unread-in-browser and open-all-unread-in-browser-and-mark-read
 with the difference of marking opened items as read or not. Both
commands are available from the feed list view and the article list
 view. Article opening is implemented in the new
open_unread_items_in_browser function of the formaction class.

To avoid mishaps, the number of articles that can be opened by these
commands is limited by the max-browser-tabs configuration command.