Use "GNU/Linux" where appropriate. Closes #381
"Appropriate" means "where we refer to a system where Linux kernel is
used in conjunction with GNU libc." Granted, GNU/Linux is not the only
example of Linux-based system, but others (e.g. Linux with musl) don't
have a nice short name, and we don't want to use an overly formal,
long-winded tone in the docs.