Separate line wrapping and whitespace consolidation
Whitespace consolidation was done in tokenize_spaced() which was used
from the wrap_lines function. To account for softwrappable lines, which
we don't want to consolidate the whitespace on, the behaviour of
tokenize_spaced what changed from consolidating whitespace to simply
tokenizing and leaving whitespace intact. Consolidation was moved to the
new consolidate_whitespace() function. Fixes #392