Fix a bug I introduced in 885760cf70c35fb632d2c872
One of the fields of `itemview_formaction` (namely, `num_lines`) wasn't
updated properly: it's supposed to contain the number of lines of text
displayed in the view, but I stored the *length* of the text there. As
a consequence, the reading progress indicator always said "Top" instead
of being updated as the user scrolls through the article.