Fix a bug that didn't render empty lines on preformatted text
Previously, we relied on the behaviour of tokenize_spaced to always
return a space token even if the string was empty to properly render
empty lines, now that consolidation was split into
consolidate_whitespace empty lines in functions that didn't call it(i.e
in HTML preformatted text) wouldn't get rendered. This commit changes
the behaviour of consolidate_whitespace to return an empty string with
an empty input and special case checks were added for empty wrappable
and softwrappable lines in the textformatter.