Fix erroneous comments in rss_feed::set_rssurl
cppcheck pointed out that I overwrite variable's value right after
assigning it, and this proved to be true. However, while looking into
it, I found that the comments I left in the function don't match the
reality as the input that it gets are a bit different from the examples
in comments. It's sheer luck that the code that I got by making
a mistake turned out to be *right*.

So I removed superfluous code, amended the comments and added a few
tests to make sure I never step into that particular trap ever again.