Make TempFile go along if temporary dir exists
Sometimes during development tests fail not as they should, but with
a SIGSEGV. In this case, temporary files won't be deleted, and so won't
be the temporary directory TempFile created. Previous version of
TempFile would complain about that and force the developer to go delete
the directory before any tests can be run. But the directory is totally
fine! (Save for a few files that now reside in it and won't be deleted
automatically.) So we now accept existing directory as normal, as long
as we can write into it.