Enable -Werror to make warnings Travis build fails
Had to silence a few warnings in auto-generated code first, of which
I am not that proud: they'll be back when we re-generate the code.
However, I think the benefits still outweigh the disadvantages, as
1) that code isn't touched very often; and 2) compiler warnings
generally mean that we're doing something fishy, and I do want to know
about that.