Build for warnings first, then for tests (#430)
[GCC's online docs on warnings]
(https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html) contain
numerous notes along the lines of "this warning is only enabled when
optimizations are enabled" and "the precision of this warning depends on
the optimization level". But optimizations are at odd with debugging.

Thus we now have two stages in our build script.

First we compile with `-O3`, hoping that this will give us the best
warnings the compiler can issue. Tests are compiled but aren't run;
we're only interested in our code not looking suspicious to the
compiler.

We then `distclean` and proceed to the second stage, where we compile
without optimizations but with profiling. This stage runs test.

As always, we only report coverage on successful build.