add support for the socks5h proxy type
The proxy type socks5h does the hostname resolution on the proxy instead
of the client (hence the name) and is therefore preferred with e.g Tor.

The socks5 type allows a client to do either with most performing the
DNS lookup locally and asking the socks5-proxy server to connect to a
specific IPv4/6. That can fail if the server has another network setup –
like no IPv6 – and in terms of Tor it means that DNS queries are leaked
and .onion addresses aren't working.