Mark item unread when it's passed to external pager
This might look like laziness, but it isn't. Even though the very same
code appears at the beginning of
itemview_formaction::process_operation(), it's really not possible to
factor it out without making bigger changes.

The block at the beginning of process_operation() is there to ensure
that no matter what commands user issues, the item that appeared in the
itemview will be marked read. This became important when things like "go
to next", "go to previous" and "go to next unread" were implemented;
since itemview isn't re-created in these cases, it was impossible to
mark article read in OP_QUIT handler. To avoid sticking
mark_unread(false) into numerous operation handlers, Andreas apparently
decided to do it every time an operation is handled (see
eca8e0c4162a759e0c46b01ef8aeebe3f790ecf8).

This means that mark_unread(false) can't be pulled further up the stack
and into view::push_itemview(). In turn, that means that for this
change, we had to copy-and-paste the code that marks the item read.

Fixes #495.