Evaluate the commit hash once when compiling
A makefile assignment with '=' will be expanded every time it is
referenced. This meant that the git command would be run every time it
was referenced in a compiler call.

Instead, it is better to use ':=' which expands the value once when it is
assigned. This prevents multiple calls to git saving a few precious CPU
cycles.