Use glob() instead of wordexp()
The code is lifted from OpenBSD patch by Aaron Bieber (@qbit)[1].

wordexp() has some security implications[2], hence it's not implemented
in OpenBSD[3]. This means our code isn't as portable as it could be. (I
don't think we're really open to any security threats here because the
config file is hardly a source of untrusted input.)

Kudos to Han Boetes for bringing this to my attention, complete with
links to the patches.

1. https://github.com/openbsd/ports/blob/0e89391b501010788b825e1c1169eb24a9be96c1/www/newsbeuter/patches/patch-src_feedhq_api_cpp
2. http://www.openwall.com/lists/oss-security/2015/02/11/3
3. http://openbsd-archive.7691.n7.nabble.com/patch-libc-wordexp-support-td159368.html