TTRSS-API: Properly escape json-strings in run_op
Up to now, the json-payload was constructed without properly escaping
the JSON strings, simply wrapping them with `""` instead.
This commit switches the payload construction to the relevant functions
from json-c, which properly handle string-escaping.

Please note that this is still not perfect, as the api still ignores the
upstream type information stated in the TTRSS-Api-docs by converting all
values to string, even bool and int parameters.