TTRSS-API: Eliminate markreadthread-class by Lambda
To mark articles read in an asynchronous fashion (without blocking the
mainloop), the TTRSS-API spawns a new thread.
To implement this thread, a new class was implemented, the
`markreadthread`, whose implementation of operator() was then used as
the tasks callable object.
This can be radically simplified by using c++11s lambdas, without
instanciating the extra class.