delete-read-articles-on-quit *marks* as deleted
This hides the articles from the user (effectively deleting them), but
still keeps the article in the cache so that future updates don't bring
them back. The articles will be deleted for real only when they fall out
of the feed.

Kudos to @polyzen for bringing this up in https://github.com/akrennmair/newsbeuter/issues/581#issuecomment-320380691