intercept browser's/pager's failure (issue #449)
newsbeuter utilizes the system(3) function to run external programs.
Examples for external programs are a text editor, an extermal pager or the
browser.
Those external programs may return a non-zero exit code. This patch intercepts
in a such case and prints an error message as well as a debugging output.