Actually create datadir
src/controller.cpp:542 does mkdir_parents(data_dir), but since the
variable is never initialized, that line did nothing. Now I initialize
it every time I find an XDG configuration, which ensures that the
datadir exists if configdir existed.

The default constructor for std::string creates empty string, so that
mkdir call is harmless if we're in dotdir mode and don't have a separate
datadir to create.