Name configuration options' args more consistently
Documentations used to use all kinds of ad-hoc names for the same
things, like command/browser-command/bookmark-command (and even "path"!)
We now do a better job at using more generic names and employing them
consistently over the entire list of configuration options.

Thanks to Stealth (on IRC) for a prod to do this.