Add hyperlinks for all config commands
The list of configuration options is rather long. In order to link to a
specific config command, turn them all to hyperlinks. This will improve
user support on IRC and the like, as one can point people directly to
the config command of interest. See also #10.

The only issue right now is that we're using old-style AsciiDoc tables
which do not support setting the vertical alignment. All table cells are
vertically centered. Thus jumping to most config command anchors scrolls
to the middle of the description as it is quite long, which is not
ideal.