Use new-style AsciiDoc tables
This allows us to use several new features. By default all table cells
are vertically top aligned, which is great for the newly inserted
configuration command hyperlinks in the previous commit. However, paragraphs will be placed in
cells which by default makes for quite a waste of vertical space. Thus we add
some CSS rules to fix this. See also #10.