Use double pipe in some DSVs to fix examples
As it turns out, some examples and default values of the configuration
commands in the documentation are incorrect, as single quotes are
interpreted as italic markers in both old- and newstyle AsciiDoc tables.
New-style tables can be configured to render certain columns as
monospace and literal. However, this does not work when a cell contains
markup – single quotes are markup.

So one solution is to use literal style columns. We also redefine the
pipe delimiter to be a double pipe delimiter. This has two advantages:
First of all we can finally fix the articlelist-format config command to
actually use pipes instead of the semicolon workaround. And secondly the
source gets a tiny bit more readable. The drawback however is, that we
must fix the CSS again, or else the literal columns are way wider than
the page, resulting in horizontal scrollbars.

Having a regular expression with negative lookahead seems quite ugly and
difficult to grasp to me, so make the code more readable by introducing
and reusing the split function. See also #10.

See also #10.