Don't repeat version number when built from Git
When Andreas Krennmair did Newsbeuter releases, he forked master branch,
generated documentation and put a tag on the result. This meant that
when git-describe was run in master, it didn't consider the tags as they
aren't part of master branch.

I do releases differently; I tag a commit in master, and release that.
But now git-describe sees those tags and includes them in result. Since
config.h prepends current version to that result, we were getting
oddities like "2.10.1-2.10.1-58-aabbcc".