Fix building of docs on FreeBSD
These sed commands fail on FreeBSD as they're interpreted as extended
regular expressions. At least, that's what the error message suggests to
me. As it turns out using the -e option was wrong in the first place as
I thought this switch enables extended regex, which it doesn't. Correct
is either -r (GNU sed) or -E (BSD sed). Both implementations support the
other's flag as well, so it shouldn't matter which we actually use.

* https://www.freebsd.org/cgi/man.cgi?query=sed&sektion=&n=1
* https://stackoverflow.com/a/3139925

Thanks to imm_ on IRC for reporting this and confirming the fix!