Fix exiting newsboat/podboat if `--help` is given
Commit a0cc1ae755386a2db342659c0a9d1fe9794f84ed introduced a bug where
supplying the `--help` option did not exit the softwares immediately as
they were supposed to do, but actually started them after showing the
help text. This commit fixes that.

Also rename the methods to indicate more clearly what they do and let
the exit code handling be done by the callers. This way we can also
return exit code 0 when `--help` was requested as most other programs do
as well. Previously, always exit code 1 was returned when the usage was
involved, even when passing `--help`.

Thanks to qgTG in IRC for reporting this bug!