Get rid of premature silent check
The main motivation behind this refactoring is to suppress getopt's
duplicate error messages whenever an unknown option is supplied.
Previously this message was generated twice, once in each of both getopt
loops.

As it turns out, enabling the silence flag for the import is not
necessary as this variable is not used anywhere before encountering the
import, so simply get rid of it.