Use more portable __linux__ define
This fixes an Alpine Linux has building newsbeuter because the distro
uses MUSL instead of glibc. To quote the original patch author:

"If you try to compile newsbeuter without -fpermissive, as ppc64le, it
will not build due to a wrong argument type.

It is wrong because it sets the proper argument if you have Linux with
Glibc. Since we use MUSL, it use the wrong argument.

This patch assure that both musl and glibc uses the same argument type."