Print `error-log` error to stderr, too
I didn't test the previous patch before committing it. Turns out it
doesn't actually show the error, because it prints it to the `view`
instance that isn't displayed yet. So I print it to stderr, too.

Ideally we'd wait for the user to read the error, or maybe abort the
startup altogether, but that's hard to do with current code.