Put the cursor at the end of the line in textfields
Pre-2.9 the cursor for qna values was set to the end of the line. This
was changed in 462745d because it caused the text before the cursor to
be hidden for some reason.

Bring back the old behaviour by calling the run() function to initialize
the field and then set the cursor to the desired position.