utils::retrieve_url: Make it possible to use cached handle
Up to now, retrieve_url has created a new handle for each request. As
connection pools and therefore reuse of existing TCP-connections is tied
to the handle in libcurl, this resulted in slow bulk updated, e.g. when
using retrieve_url in an remote_api, where all requests target the same
API/host.