Use unique_ptr::reset() to delete the object
release(), which we previously used, only asks unique_ptr to stop
managing the pointer, but doesn't ask for destructor to be called. This
creates a memory leak; it's a minor problem since it's just
a short-lived unit test, but since I already noticed it, I might as well
fix it.