Ignore return value of setlocale()
setlocale() has been failing for some MacOS users because they didn't
have LANG environment variable (or had it set to a non-existent locale).
Analysis by Jacob Wahlgren has shown that we're actually calling
setlocale() from libintl, not the one from stdlib, and it has some
surprising behaviour on MacOS and Windows.

Furthermore, it turned out that all other projects that we checked
ignore the return value of setlocale(). If Mutt and Vim do it, why not
Newsboat, too?

Closes #156.