Resolve AsciiDoc's {backslash} in example config
Previously, the generated example config file contained verbatim the
AsciiDoc's {backslash} rule as defined in the DSV source file:

    +{backslash}"+

This is not very human-readable. First, there is this strange backslash
word surrounded by curly braces and second, the whole thing is put in
plus signs instead of grave accents, which most people are used to.
Thus, make it more reader-friendly.

Also use ` instead of + for code markup, where possible. The plus sign
allows for attribute resolution, but the grave accent does not. This
also helps the generated example config to be more comprehensible.