Avoid TOCTOU in mkdir_parents
"TOCTOU" means "time of check, time of use". We check if the directory
exists, and if not, we create it. The problem is that someone else
might've created the directory while we were in between calls to
access() and mkdir(). To handle this, we examine the return code of
mkdir() and don't consider it an error if the directory we're creating
already exists.

Bug found by Coverity Scan.