Add support for sorting feedlist on recent updates
When using the `sort' and `revsort' actions in the feedlist, added a new
option (bound to the `r' key for `recent') which sorts the view based on
the most recent publication date for each feed.

This allows a user to order their feeds based on recent activity, as
feeds that have not seen an update in considerable time will naturally
sink to the bottom of the list (or top, if revsorting).