Fix bind-key accepting invalid operations
While binding a key using bind-key we didn't verify the return value of
get_opcode to make sure we're binding to a valid operation.

This makes newsboat throw an error when trying to bind to an invalid op
similarly to what happens when trying to use an invalid operation in a
macro.