Fix podboat segfault when parsing comment in queue
Podboat does the initial parsing in queues via `tokenize_quoted()` which
returns an empty vector if it comes across a comment, which was never
actually checked in the code and resulted in a segfault.

Since podboat re-writes the file on exit we can not preserve the
comments without a lot of refactoring, so as a middle ground solution
issue an interactive warning to the user instead.

Thanks to hseg on IRC for the discovering this!