Fix 0..9 keys in article view
In article view, numeric keys open a URL with a given number. I broke
that behaviour in 485082f43f49f1c30974621107b1fa197618bf67 because
I didn't fill articleview's `links` field, so the view thought that
there are no links to open.

Thanks to @j605 for reporting this on IRC!