Fix failing test for utils::resolve_tilde
That test didn't, and couldn't, pass because `resolve_tilde` works
around the missing HOME variable by checking /etc/passwd. There is no
realistic way to intercept or mock that, so I opted to remove the test
entirely.

This test case made it through CI because Travis configuration was
wrong: when C++ tests failed, it didn't fail the build, proceeding to
Rust tests instead. I fixed it with some shell magic that will run both
suites and fail the build if *any* of them fail.

Fixes #369.