Initialize global logger lazily
This saves us from one possible mistake, and it doesn't affect
performance at all, so why not do it?

(Logger is held within a OnceCell, which uses an atomic byte to keep
track of whether the cell was initialized or not. That atomic is checked
on each access. Thus, "lazy" initialization is just as costly as
"static" one, so we can choose whichever suits us best.)