Remove SIGCHLD handler
It was only needed for functions that used fork(): run_program(),
run_command(), and get_output(). Now that all of these are ported to
Rust, we don't need the handler to clean up after children. Furthermore,
we actively don't *want* that handler, since it prevents Rust's
std::process::Command from working correctly. For details on that, see
https://github.com/newsboat/newsboat/pull/358#issuecomment-438782970