Partially revert "Escape URLs for Inoreader, Old Reader"
This partially reverts commit bf67b204ef623f4ed236ee1467d1bcd90351fb58.

That commit assumed that Inoreader and The Old Reader implement the
Google Reader API the same way, but that assumption turned out to be
wrong.

Inoreader uses feeds' URLs as their identifiers, with all the slashes
and ampersands and question marks. Thus, calls to Inoreader need
URL-escaping, to prevent our query parameters from mixing with
parameters in the feed identifier. Inoreader doesn't care that we escape
a slash between "feed" and "feed identifier" parts: apparently it uses
everything that follows the API prefix as an identifier.

The Old Reader, on the other hand, assigns each feed a hexadecimal
string, and expects us to keep a slash between "feed" and "feed
identifier". That's the contract that we broke, and which is mended by
this commit.

Fixes #406.