Add format option for feed titles in query-feeds
Introduces an '%N' formatting option to be used in the 'download-path'
and 'download-filename-format' format strings, which exposes the
items original feed-title, even if the item is selected through
a query-feed.

This allows users to group feeds under a query-feed in newsboat, while
maintaining seperate directories and naming-conventions in the filesystem.

If the item is not selected through a query feed, '%N' and '%n' are
exactly the same.