Make DELETE_ALL work in query feeds, too (#456)
The problem was that in query feeds, feed->rssurl() returns the query.
Since none of the items in the database have the query as their
"rssurl", no items will be marked as deleted.

This commit checks if the feed is a query feed, and if so, marks items
deleted one by one.