Tidy up numeric tests for fmt!
There is no point in checking min/max values in Rust because we know the
exact size of each type. This also removes the problem with
std::i64::MAX literal, so I changed the test to test that value rather
than MAX-1 (which is what C++ tests).

Since long int and long long int are both 64 bits on x86_64, their tests
were merged into a single i64 test. Ditto for unsigned and u64.