Test pointer formatting in platform-independent way
"%p" format specifier formats pointers in an implementation-defined way,
so we can't simply compare it to "(null)" or whatever. What we do
instead is:
1. check that two pointers of different types (void* and int*) pointing
   to the same object (an int) are formatted the same;
2. all null pointers are formatted the same, regardless of their type.

I couldn't think of any other properties to check, and we don't even use
%p formatter, so the point is kind of mute. I just wanted to have tests
in place, since we do allow passing pointers into the formatting
function.