Make tests pass for OP_SAVEALL
This operation is not bound to anything by default, so tests that check
each operation failed. Where possible, I changed the tests to compare
new KeyMap to the default one—adding more operations won't break those
test anymore.

One test couldn't be fixed in this way; it'll keep failing as new
operations are added. Will have to live with that for now, I guess.