If using sections, put leaf REQUIREs into them
We often use SECTIONs for conditional setup, e.g.:

```c++
RssFeed f;
SECTION("Title is FOO") {
    f.set_title("FOO");
}
SECTION("Title is BARBAZ") {
    f.set_title("BARBAZ");
}
REQUIRE(f.title().length() == 3);
```

When such a test fails, it's impossible to tell which SECTION was
executed on this particular run. This makes it very hard to debug.

The proper way to do it is to run the checks from SECTIONS. I often do
this with lambdas: I put checks in a lambda, and run the lambda as
a last statement in a section.

Some actions, like cleanup, genuinely need to be ran *after* all
sections, but since we rely on RAII most of the time, such code is rare
in our tests.