Fix broken `pipe-to` default key in table
Previously the default key column was empty for the `pipe-to` operation,
however its description column started with a literal pipe character,
immediately followed by the description text. This is caused by the use
of double pipes to delimit table cells: in this case the default key
binding pipe was interpreted as part of the first of the two cell
separator pipes, leaving the last pipe as part of the third column.

Escaping the default key pipe with a backslash does not work as this
will put three pipe characters followed by the description in the
default key column. The rest of the table is then messed up by this drop
of the table cell (the `sort` operation is in the description column).

Using a space after the the default key binding `|` fixes the problem as
the table cell parser now treats the last two double pipe characters as
the cell delimiters.