Strip comments before the backticks are evaluated
Since the backticks should be evaluated before tokenizing
(see https://github.com/newsboat/newsboat/pull/493 for discussion),
the comments need to be stripped even before that, so that backticks
can be safely used for quoting.
Not doing so can cause various unexpected issues
(such as having "`yes`" in commented line causing newsboat to crash
with out-of-memory errors, as the `yes` command produces infinite output).