Fix incorrect setup of one ConfigPaths test
The comment says that the XDG data dir shouldn't be created, but the
very next line created it. The test expected create_dirs() to fail
afterwards, and it did--but for an entirely different reason:
- XDG config dir doesn't exist, so ConfigPaths decides to use dotdir
- the home directory was made read-only
- create_dirs() failed to create the dotdir, and returned false

The presence of XDG data dir turns out to be irrelevant.

This commit fixes both C++ and Rust tests, by creating the XDG config
dir, and *not* creating the XDG data dir. The outcome of the test is the
same, but now it tests what we want it to test.