OP_SAVEALL: account for repeating article filenames
Our previous approach generated a filename for each article, and checke
if that filename existed in the filesystem. But if all articles had the
same name, which didn't exist in the filesystem, our approach reported
that there are no conflicts--yet asked for an overwrite every time.

To fix this, we now check how many *unique* filenames we generated, and
also check how many of them are already present on the filesystem.

It's still possible for our count to be inaccurate, because more files
could've been created between us counting and us writing something. But
this should be a relatively unlikely case, and I don't want to re-run
calculation every time we run into a conflict.

Fixes #591.