Drop TODO list
These days, our issue tracker is our TODO, so this file is unnecessary.
My resolutions about its items are:

> in the article view, the headers should be an own color configuration
> element

Added to the list in 2009 (66db6d0e73eac640a228bfbb36479c03fac92a26),
and never asked for again. Probably not important to anyone. Can
reconsider if someone asks for it.

> when reload is finished, the focus shall be in the current form's list

I believe we already have that.

> implement RFC 5005 (http://www.ietf.org/rfc/rfc5005.txt)

Interesting. Added an issue about it: #628

> implement a single key with which it is possible to go through all
> unread articles (like space in slrn)

That's what "n" key does. Not sure if I'm missing something, because
I never used slrn. We can discuss this further if someone asks for this
feature.

> make it configurable whether "n" jumps to the next feed or stays in
> the same

`goto-next-feed` setting controls that now.