Give mock Matchable in Matcher tests a unique name
This is a precaution against one-definition rule violations: someone
might need a mock Matchable in another module, and they might just copy
the one we had here. They won't change the name of the struct, because
it's a nice generic `testMatchable`, and boom---ODR violated. When
definitions of the structs diverge, tests will start failing, and the
reason won't be obvious at all.

I should know, because I did exactly that.

Renaming the struct should help, because people will itch to rename it
in order to match the class under test. At least I hope they do.