Rewrite utils::resolve_tilde to operate on PathBuf
Everything's easier when proper types are used: the body became much
simpler, and now it's also easier to use this function with the rest of
Rust standard library. This isn't a theoretical concern: my next commit
would've been more complicated if I didn't rewrite this function first.

One of the tests had to be adjusted: Rust always appends a slash at the
end of the home directory path. This is safe change, because additional
slashes do no harm, both in Rust and in C++.