Restore formatting for options in manpages
This removes a single space character before each option. Apparently,
Asciidoc stripped it, but Asciidoctor preserves it, so all the options
got mis-aligned a bit.

This also uses _underscores_ in place of `quotes`. This didn't affect
the tables in HTML, but in manpages, it underscores the options (just
like they were previously).