Remove Coverity badge from the README
Submitting builds there was always troublesome, so I never did it
regularly. The analyzer is falling behind the times, and now requires
booting Linux with a special parameter.

It did found a number of bugs in our code, but the pain of maintaining
it is not worth it for me.