Make MatcherMockMatchable configurable
Previous version hurt readability, because attributes were defined in
separately from the tests. To understand a test, one had to read both
the definition of MatcherMockMatchable, and the test itself.

Previous version also discouraged writing tests with custom values; it
simply did not support that. Yet this is exactly what I need to test
regexes, hence this commit.