Make our Docker explanation friendlier to users
Now that it's exposed to end-users, it pays to streamline it further.
Instead of cross-compilation, the guide now showcases a more mundane
native compilation. The Dockerfile used for that got defaults for all
settings, so `docker build` and `docker run` are as simple as possible.

Dockerfiles now have comments explaining what the images are for, and
how to build and run them. I doubt any user will actually go there, but
it might help some contributors.