RegexManager: Update tests because stfl tags overwrite (no nesting)
My assumption was that the following line
`The <0>quick <1>brown</> fox`
would be rendered with ` fox` having the `<0>` style.
With some experimentation, I found out stfl does not work like that.

Instead, stfl works by overwriting the style whenever it encounters a
tag and `</>` will directly return to the default style (`style_normal`).