Matcher: make getters return strings, not refs
This doesn't make much difference right now, but when the guts of
Matcher are ported to Rust, it'd be awkward to return a `const
std::string&`, since all members of C++ Matcher will be gone, and we
won't be able to return a temporary anymore. Better to forestall this
problem by changing the interface.