Split test/test-helpers.h into separate files
test-helpers header grew too large. Many test units depend on it, so
changing something in test-helpers.h resulted in re-compiling half the
test suite -- even the parts that don't use the changed helper.

I don't think there is a diff-friendly way to make these changes, so
they're all in one commit. Sorry about that.

The only noteworthy thing here is that the contents of
test/test-helpers.cpp were moved to test/test-helpers/envvar.cpp. This
means that envvar.cpp now houses *both* the implementation for EnvVar
and its tests. We can do this because ultimately, all of this code is
linked into the test executable, so we don't need to do anything special
for our TEST_CASEs to be run.