Pass up,down,pageup,pagedown,home,end to STFL
These operations were ignored When encountered in macros.
For regular keypresses, the actions were already handled by STFL
(configured using STFL variables like `bind_page_up`).

We call `get_form()->run(0)` to let STFL directly process the input.
This is needed to ensure that the commands of a macro are executed in
the correct order.