Allow binding TAB key to operations
Normally, STFL switches focus between widgets when TAB is pressed.
It does not pass on the TAB event so we cannot bind any actions to it.

By registering an `on_TAB` handler, STFL directly returns the TAB event
without processing it.

Most of our STFL forms only have one widget which can be focussed.
For those forms, removing the STFL internal TAB handling does not change
the behavior.
The only exception is the `filebrowser.stfl` form (used by `FileBrowser`
and `DirBrowser`). That form allowed switching between the directory
listing and the filename at the bottom.