Add Rust implementation of `FilterParser`
This introduces a breaking change: parse tree now contains information
about the type of the right-hand side argument of comparison operator.
The C++ implementation stored them as strings.

Even though it's an internal change, it will be exposed to the users,
since the C++ `Matcher` had to work around this type erasure somehow and
created some unwanted behaviours. For a detailed list, please see the
commit that adds the Rust implementation of `Matcher`.