Add Rust implementation of `Matcher`
This assumes that users never compare string attributes to numbers, e.g.
they don't do `title = 42`. This lets us parse strings as numbers and
`unwrap()` fearlessly. That's a regression that will be fixed once
I extended the C++ test suite with checks for error cases.

As mentioned in the commit that adds a Rust implementation of
`FilterParser`, expressions' parse tree now stores information about the
type of right-hand side argument. This enables us to fix a number of
unwanted behaviours that were present in the C++ `Matcher`:

1. **equality worked by comparing strings**. As a result, filter
   expressions like `age = 0042` never matched anything, since `age`
   attribute is a number, and Newsboat never formats numbers with
   leading zeroes.

   New matcher takes the types into account, thus leading zeroes in
   numbers are now ignored, and equality comparison with numbers
   actually performs a numeric comparison;

2. **numbers could be used as regular expressions**.

   New matcher always returns `false` for cases like `attr =~ 42`, and
   `true` for `attr !~ 42`.

3. **ordering relations used numeric prefixes of the string**. E.g. an
   expression `age > "7 days"` worked fine, and even did what users
   expected, since `age` is a number measured in days. However, `age > "7
   weeks"` *didn't* work as expected, comparing `age` to 7 days instead.

   New matcher always returns `true` for `attr > "string"` and `attr >=
   "string"`, and `false` for `attr < "string"` and `attr <= "string"`.