Fix "feeddate" attr containing fixed string
Before it was broken in 48a9fdc1a37bf5eff1c35fa168f372f4258f31ef this
attribute used %c format, which is locale-dependent. This commit uses
the same format as used in "date" attribute of an article. It's a bit
more portable, but it's also marked for translation, so we're back to
square one. I guess the expectation is for users to look at their
screens, see what format is used in the itemview, and use the same in
their filters.