Docs: stop listing full paths to files
Since Newsboat and Podboat can store their files either in the dotdir or
in XDG dirs, it's misleading to list only the dotdirs paths. In most
places, full paths were replaced by the names of the files, plus a note
of where to find them relative to the urls and cache files. In the
configuration commands table, we list both possible paths.

This approach is less wordy than listing both paths all the time, but
users will have to look through the doc and find out where Newsboat
stores the files. Given the "First Steps" section already mentions both
possible paths, I think it's acceptable to be less wordy in other
places. The config commands table is an exception because it's
a reference which is meant to be read in pieces, so all the knowledge in
it should be "local", without "external" references.

Addresses https://github.com/newsboat/newsboat/pull/907#issuecomment-619124517.